pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sleeplab-format"
dynamic = ["version"]
description = "An interoperable and harmonized format for polysomnography recordings."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Riku Huttunen", email = "riku.huttunen@uef.fi" },
]
dependencies = [
    "numpy",
    "pandas",
    "pyarrow",
    "pyyaml",
    "pydantic>=2.0",
    "scipy",
    "zarr"
]

[project.scripts]
slf-extract = "sleeplab_format.extractor.cli:run_cli"

[project.urls]
Documentation = "https://github.com/UEF-SmartSleepLab/sleeplab-format#readme"
Issues = "https://github.com/UEF-SmartSleepLab/sleeplab-format/issues"
Source = "https://github.com/UEF-SmartSleepLab/sleeplab-format"

[tool.hatch.version]
path = "src/sleeplab_format/version.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]

[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=src/sleeplab_format --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["39", "310", "311"]

[tool.hatch.envs.docs]
dependencies = [
  "mkdocs-material",
  "mkdocstrings[python]>=0.18"
]

[tool.hatch.envs.docs.scripts]
build = "mkdocs build"
serve = "mkdocs serve"
publish = "mkdocs gh-deploy --force"

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "src/sleeplab_format/version.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
